+++ packages/container/examples/basic/index.widget.js
@@ -1,4 +1,5 @@
-export default () => {
+export default a => {
+  console.log(a.data);
   let element;
   return {
     async bootstrap() {