+++ packages/container/examples/child-widget/index.widget.js
@@ -8,7 +8,7 @@ export default () => {
       console.log('demo bootstrap');
       element = document.createElement('div');
     },
-    async mount({ container, context }) {
+    async mount({ container }) {
       console.log('demo mount');
 
       element.innerHTML = `
@@ -23,7 +23,7 @@ export default () => {
       container.appendChild(element);
 
       container.querySelector('#close-self').onclick = () => {
-        context.unmount();
+        container.unmount();
       };
 
       container.querySelector('#load-remote-widget').onclick = () => {