+++ packages/container/src/WebWidgetDependencies.js
@@ -1,22 +1,5 @@
 /* global window */
 
-const CONTEXT = Symbol('context');
-const CREATE_PORTAL = Symbol('createPortal');
-
-function createContext(view) {
-  return {
-    mount() {
-      return view.mount();
-    },
-    update(properties) {
-      return view.update(properties);
-    },
-    unmount() {
-      return view.unmount();
-    }
-  };
-}
-
 /**
  * @summary Web Widget Dependency Injection
  */
@@ -40,72 +23,17 @@ export class WebWidgetDependencies {
    */
   get container() {
     const view = this.ownerElement;
-    view.renderRoot = view.renderRoot || view.createRenderRoot();
+
+    if (!view.renderRoot) {
+      view.renderRoot = view.createRenderRoot();
+      ['mount', 'update', 'unmount'].forEach(name => {
+        if (!view.renderRoot[name]) {
+          view.renderRoot[name] = properties => view[name](properties);
+        }
+      });
+    }
 
     return view.renderRoot;
-  }
-
-  /**
-   * Container context interface
-   * @type {object}
-   */
-  get context() {
-    if (!this[CONTEXT]) {
-      this[CONTEXT] = createContext(this.ownerElement);
-    }
-
-    return this[CONTEXT];
-  }
-
-  /**
-   * Create a portal
-   * @function
-   * @param {HTMLWebWidgetElement}  webWidgetElement Web widget container
-   * @param {string}                name             Destination name
-   */
-  get createPortal() {
-    if (!this[CREATE_PORTAL]) {
-      this[CREATE_PORTAL] = (widget, name) => {
-        const view = this.ownerElement;
-        const HTMLWebWidgetElement = view.constructor;
-
-        const factory = HTMLWebWidgetElement.portalDestinations.get(name);
-        const portal = factory ? factory() : null;
-
-        if (!portal) {
-          throw new Error(`The portal cannot be found: ${name}`);
-        }
-
-        if (!(portal instanceof HTMLWebWidgetElement)) {
-          throw new Error(
-            `Portal must be an instance of "HTMLWebWidgetElement": ${name}`
-          );
-        }
-
-        if (this.sandboxed && !widget.isConnected) {
-          // Start the sandbox in the current scope
-          this.container.appendChild(widget);
-        }
-
-        if (!widget.slot) {
-          widget.slot = '';
-        }
-
-        const oldWidget = portal.querySelector(`[slot="${widget.slot}"]`);
-
-        if (oldWidget) {
-          portal.removeChild(oldWidget);
-        }
-
-        portal.appendChild(widget);
-        portal.mount();
-        view.portals.push(portal);
-
-        return createContext(portal);
-      };
-    }
-
-    return this[CREATE_PORTAL];
   }
 
   /**
@@ -113,8 +41,7 @@ export class WebWidgetDependencies {
    * @type {(object|array)}
    */
   get data() {
-    const data = this.ownerElement.data;
-    return Array.isArray(data) ? [...data] : { ...data };
+    return this.ownerElement.data;
   }
 
   set data(value) {
@@ -122,34 +49,17 @@ export class WebWidgetDependencies {
   }
 
   /**
-   * Application name
-   * @type {string}
-   */
-  get name() {
-    return this.ownerElement.name;
-  }
-
-  /**
-   * Application parameters
+   * Application environment variables
    * @type {object}
    */
-  get parameters() {
+  get env() {
     return [...this.ownerElement.attributes].reduce(
       (accumulator, { name, value }) => {
         accumulator[name] = value;
         return accumulator;
       },
-      {}
+      Object.create(null)
     );
-  }
-
-  /**
-   * Sandbox mode
-   * @type {boolean}
-   */
-  get sandboxed() {
-    const { sandboxed } = this.ownerElement;
-    return sandboxed;
   }
 }