+++ packages/system-loader/test/index.test.js
@@ -1,15 +1,6 @@
 import 'systemjs';
 import { HTMLWebWidgetElement } from '@web-widget/container';
-import '@web-widget/sandbox';
 import '../src/index.js';
-
-const get = url =>
-  fetch(url).then(res => {
-    if (!res.ok) {
-      throw Error([res.status, res.statusText, url].join(', '));
-    }
-    return res.text();
-  });
 
 const dataUpdate = widget =>
   new Promise(resolve => {
@@ -26,30 +17,12 @@ const dataUpdate = widget =>
     widget.mount();
   });
 
-const srcCase = async sandboxed => {
+const srcCase = async () => {
   const widget = document.createElement('web-widget');
   widget.inactive = true;
   widget.name = 'TestWidget';
   widget.type = 'system';
-  widget.sandboxed = sandboxed;
   widget.src = '/test/test.widget.js';
-  document.body.appendChild(widget);
-
-  await widget.mount();
-  await dataUpdate(widget);
-
-  if (widget.data.lifecycle !== 'mount') {
-    throw new Error(`Mount error`);
-  }
-};
-
-const textCase = async sandboxed => {
-  const widget = document.createElement('web-widget');
-  widget.inactive = true;
-  widget.name = 'TestWidget';
-  widget.type = 'system';
-  widget.sandboxed = sandboxed;
-  widget.text = await get('/test/test.widget.js');
   document.body.appendChild(widget);
 
   await widget.mount();
@@ -62,13 +35,4 @@ const textCase = async sandboxed => {
 
 describe('Load module', () => {
   it('Load the System module', async () => srcCase());
-
-  it('Load the System module: local', async () => textCase());
-});
-
-describe('Sandbox mode', () => {
-  it('Load the System module', async () => srcCase(true));
-
-  // 由于使用了 data: 协议，但是 WebSandbox 无法支持它
-  // it('Load the System module: local', async () => textCase(true));
 });